Move get_child_pid test to Testing.Pipeline test
